feat(backend): add Semaphore wrapper reads, view service, and vote/semaphore endpoints

+++ evote-backend/src/main/java/com/evote/backend/dto/SubmitVoteResponse.java
@@ -3,12 +3,14 @@ package com.evote.backend.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 public class SubmitVoteResponse {
     private String transactionHash;
     private VoteStatus status;
     private Long timestamp;
-    private Long blockNumber;
+    private BigInteger blockNumber;
     private String message;
 }
